keep scene3d() in colzette.py and delete deprecated_or_unused.py + archive/
