--- archive/data/climate/IDEEV_2023_INRAE_STATION_91272001.csv
@@ -1,134 +1,0 @@
-Nom de la demande : Moulon_2023
-Pas de temps : Journalier
-Date de début : 01/09/2023
-Date de fin : 31/12/2023
-Station(s) : 91272001
-
-TN : Température minimale (Degrés Celsius)
-TX : Température maximale (Degrés Celsius)
-PAR : Rayonnement PAR (Joules / cm²)
-RG : Rayonnement global (Joules / cm²)
-
-NUM_POSTE;AN;MOIS;JOUR;TN;TX;PAR;RG;
-91272001;2023;9;1;16.1;23.3;414.0;870;
-91272001;2023;9;2;13.6;28.8;818.0;1762;
-91272001;2023;9;3;17.1;30.7;936.0;2072;
-91272001;2023;9;4;15.1;30.9;973.0;2197;
-91272001;2023;9;5;18.1;33.2;846.0;1913;
-91272001;2023;9;6;17.3;33.0;865.0;1964;
-91272001;2023;9;7;14.1;32.9;909.0;2009;
-91272001;2023;9;8;16.8;34.2;881.0;1915;
-91272001;2023;9;9;18.0;34.5;859.0;1894;
-91272001;2023;9;10;21.5;33.9;839.0;1831;
-91272001;2023;9;11;17.5;27.4;612.0;1336;
-91272001;2023;9;12;16.7;18.6;205.0;411;
-91272001;2023;9;13;16.1;22.8;518.0;1126;
-91272001;2023;9;14;12.8;24.3;723.0;1603;
-91272001;2023;9;15;12.9;27.2;859.0;1929;
-91272001;2023;9;16;13.0;25.2;572.0;1208;
-91272001;2023;9;17;14.8;27.1;457.0;975;
-91272001;2023;9;18;16.4;23.1;605.0;1309;
-91272001;2023;9;19;12.0;20.5;618.0;1315;
-91272001;2023;9;20;15.1;23.8;635.0;1380;
-91272001;2023;9;21;14.0;16.6;215.0;456;
-91272001;2023;9;22;11.4;16.4;530.0;1179;
-91272001;2023;9;23;8.4;18.1;520.0;1136;
-91272001;2023;9;24;6.7;20.3;797.0;1724;
-91272001;2023;9;25;9.0;23.0;629.0;1342;
-91272001;2023;9;26;11.9;24.9;619.0;1347;
-91272001;2023;9;27;11.9;26.1;552.0;1196;
-91272001;2023;9;28;14.7;21.9;612.0;1337;
-91272001;2023;9;29;16.9;20.7;473.0;1056;
-91272001;2023;9;30;8.9;21.9;577.0;1266;
-91272001;2023;10;1;10.8;27.0;688.0;1488;
-91272001;2023;10;2;12.4;28.6;676.0;1480;
-91272001;2023;10;3;15.0;19.5;399.0;884;
-91272001;2023;10;4;7.3;17.8;540.0;1194;
-91272001;2023;10;5;8.5;19.2;514.0;1117;
-91272001;2023;10;6;6.4;21.6;658.0;1451;
-91272001;2023;10;7;9.2;23.6;620.0;1336;
-91272001;2023;10;8;9.6;26.6;607.0;1318;
-91272001;2023;10;9;11.4;25.7;544.0;1184;
-91272001;2023;10;10;10.9;26.5;541.0;1190;
-91272001;2023;10;11;12.9;24.4;585.0;1279;
-91272001;2023;10;12;14.5;21.6;305.0;640;
-91272001;2023;10;13;17.4;24.0;403.0;860;
-91272001;2023;10;14;9.1;13.9;346.0;741;
-91272001;2023;10;15;6.0;12.6;504.0;1100;
-91272001;2023;10;16;2.2;13.1;462.0;1041;
-91272001;2023;10;17;2.5;15.6;465.0;981;
-91272001;2023;10;18;10.0;18.5;170.0;343;
-91272001;2023;10;19;14.5;21.4;384.0;806;
-91272001;2023;10;20;12.7;18.1;308.0;658;
-91272001;2023;10;21;11.1;14.4;232.0;491;
-91272001;2023;10;22;10.3;15.7;376.0;808;
-91272001;2023;10;23;5.8;11.4;84.0;169;
-91272001;2023;10;24;10.3;15.2;358.0;999;
-91272001;2023;10;25;10.3;15.1;232.0;519;
-91272001;2023;10;26;10.3;14.0;203.0;433;
-91272001;2023;10;27;9.7;14.0;231.0;498;
-91272001;2023;10;28;10.4;14.5;116.0;243;
-91272001;2023;10;29;10.6;15.9;304.0;663;
-91272001;2023;10;30;9.3;12.3;99.0;207;
-91272001;2023;10;31;7.8;14.6;245.0;537;
-91272001;2023;11;1;10.5;13.8;154.0;332;
-91272001;2023;11;2;8.7;12.9;396.0;871;
-91272001;2023;11;3;4.6;9.2;183.0;400;
-91272001;2023;11;4;7.6;13.9;221.0;481;
-91272001;2023;11;5;9.5;13.7;224.0;487;
-91272001;2023;11;6;7.5;12.3;254.0;545;
-91272001;2023;11;7;6.1;12.0;338.0;728;
-91272001;2023;11;8;6.0;12.1;146.0;313;
-91272001;2023;11;9;6.8;11.1;175.0;373;
-91272001;2023;11;10;5.5;10.7;194.0;418;
-91272001;2023;11;11;4.4;11.8;251.0;546;
-91272001;2023;11;12;6.0;14.1;100.0;197;
-91272001;2023;11;13;11.2;16.4;71.0;142;
-91272001;2023;11;14;10.8;12.5;83.0;169;
-91272001;2023;11;15;8.1;12.6;232.0;485;
-91272001;2023;11;16;7.5;12.0;84.0;168;
-91272001;2023;11;17;6.1;11.8;257.0;534;
-91272001;2023;11;18;8.1;14.1;37.0;70;
-91272001;2023;11;19;11.0;14.2;260.0;548;
-91272001;2023;11;20;8.8;11.3;174.0;367;
-91272001;2023;11;21;7.9;11.7;112.0;229;
-91272001;2023;11;22;4.8;9.1;299.0;618;
-91272001;2023;11;23;5.3;11.1;82.0;171;
-91272001;2023;11;24;6.3;10.4;119.0;257;
-91272001;2023;11;25;1.3;7.6;242.0;507;
-91272001;2023;11;26;-1.0;6.1;220.0;479;
-91272001;2023;11;27;2.9;8.5;35.0;70;
-91272001;2023;11;28;2.4;6.5;209.0;454;
-91272001;2023;11;29;0.9;8.5;237.0;498;
-91272001;2023;11;30;1.3;5.7;163.0;339;
-91272001;2023;12;1;0.3;3.5;90.0;195;
-91272001;2023;12;2;-2.9;2.9;249.0;617;
-91272001;2023;12;3;-1.9;3.0;104.0;223;
-91272001;2023;12;4;2.0;9.0;75.0;157;
-91272001;2023;12;5;5.5;7.8;74.0;154;
-91272001;2023;12;6;2.7;8.6;198.0;412;
-91272001;2023;12;7;1.7;6.9;92.0;192;
-91272001;2023;12;8;5.9;11.1;166.0;362;
-91272001;2023;12;9;5.4;12.8;43.0;83;
-91272001;2023;12;10;7.6;12.5;112.0;241;
-91272001;2023;12;11;9.3;11.4;138.0;291;
-91272001;2023;12;12;9.6;12.3;99.0;215;
-91272001;2023;12;13;6.3;8.9;93.0;200;
-91272001;2023;12;14;6.4;7.5;111.0;236;
-91272001;2023;12;15;6.5;9.6;120.0;260;
-91272001;2023;12;16;2.4;11.3;241.0;482;
-91272001;2023;12;17;-0.8;8.0;253.0;528;
-91272001;2023;12;18;-0.9;4.0;58.0;127;
-91272001;2023;12;19;2.4;9.0;83.0;174;
-91272001;2023;12;20;3.6;9.4;156.0;318;
-91272001;2023;12;21;8.0;11.1;54.0;110;
-91272001;2023;12;22;8.3;10.5;87.0;178;
-91272001;2023;12;23;7.9;10.9;78.0;165;
-91272001;2023;12;24;8.6;11.4;139.0;290;
-91272001;2023;12;25;10.1;11.2;68.0;142;
-91272001;2023;12;26;8.8;9.9;55.0;114;
-91272001;2023;12;27;6.6;9.9;167.0;364;
-91272001;2023;12;28;8.4;10.4;98.0;211;
-91272001;2023;12;29;8.3;10.8;82.0;175;
-91272001;2023;12;30;6.7;10.4;81.0;170;
-91272001;2023;12;31;6.8;10.0;215.0;455;
