--- archive/data/parameters/global_params_fababean.csv
@@ -1,34 +1,0 @@
-"Species"	"Parameter"	"Level"	"Value"
-"Fababean"	"coeff_width_leaflet1"	"Default"	0.690536840217465
-"Fababean"	"coeff_width_leaflet1"	"q1"	0.642424242424242
-"Fababean"	"coeff_width_leaflet1"	"q3"	0.744055944055944
-"Fababean"	"coeff_width_leaflet2"	"Default"	0.52576180667712
-"Fababean"	"coeff_width_leaflet2"	"q1"	0.5
-"Fababean"	"coeff_width_leaflet2"	"q3"	0.565217391304348
-"Fababean"	"coeff_petiole_leaflet1"	"Default"	0.233540714011381
-"Fababean"	"coeff_petiole_leaflet1"	"q1"	0.181445603576751
-"Fababean"	"coeff_petiole_leaflet1"	"q3"	0.324127264692886
-"Fababean"	"coeff_petiole_leaflet2"	"Default"	0.920480222899129
-"Fababean"	"coeff_petiole_leaflet2"	"q1"	0.819672131147541
-"Fababean"	"coeff_petiole_leaflet2"	"q3"	1.05660377358491
-"Fababean"	"ins_angle"	"Default"	37.7985074626866
-"Fababean"	"ins_angle"	"q1"	26.25
-"Fababean"	"ins_angle"	"q3"	45
-"Fababean"	"rmax"	"Default"	0.65005192617206
-"Fababean"	"rmax"	"q1"	0.583699525398067
-"Fababean"	"rmax"	"q3"	0.746347200381635
-"Fababean"	"k"	"Default"	2.62971808931429
-"Fababean"	"k"	"q1"	1.16485781648376
-"Fababean"	"k"	"q3"	2.49186625645388
-"Fababean"	"L_height"	"Default"	77.4384869869446
-"Fababean"	"L_height"	"q1"	29
-"Fababean"	"L_height"	"q3"	40
-"Fababean"	"x0_height"	"Default"	538.021521784714
-"Fababean"	"x0_height"	"q1"	400
-"Fababean"	"x0_height"	"q3"	600
-"Fababean"	"k_height"	"Default"	0.0207298398070234
-"Fababean"	"k_height"	"q1"	0.005
-"Fababean"	"k_height"	"q3"	0.1
-"Fababean"	"phyllot"	"Default"	176.745283018868
-"Fababean"	"phyllot"	"q1"	159
-"Fababean"	"phyllot"	"q3"	192
