--- archive/data/parameters/global_params_rapeseed.csv
@@ -1,22 +1,0 @@
-"Species"	"Parameter"	"Level"	"Value"
-"Rapeseed"	"coeff_width_leaf"	"Default"	0.415592882407371
-"Rapeseed"	"coeff_width_leaf"	"q1"	0.37296365914787
-"Rapeseed"	"coeff_width_leaf"	"q3"	0.464209515681673
-"Rapeseed"	"coeff_petiole_leaf"	"Default"	0.351519470899816
-"Rapeseed"	"coeff_petiole_leaf"	"q1"	0.27816573556797
-"Rapeseed"	"coeff_petiole_leaf"	"q3"	0.435530514559234
-"Rapeseed"	"ins_angle"	"Default"	34.6415662650602
-"Rapeseed"	"ins_angle"	"q1"	20
-"Rapeseed"	"ins_angle"	"q3"	43.75
-"Rapeseed"	"rmax"	"Default"	0.619420312123767
-"Rapeseed"	"rmax"	"q1"	0.559365582556123
-"Rapeseed"	"rmax"	"q3"	0.661634100214167
-"Rapeseed"	"k"	"Default"	7.42771083455965
-"Rapeseed"	"k"	"q1"	3.27769994094298
-"Rapeseed"	"k"	"q3"	10.3479955061558
-"Rapeseed"	"growth_node"	"Default"	0.000342993585686865
-"Rapeseed"	"growth_node"	"q1"	0.000198487317705069
-"Rapeseed"	"growth_node"	"q3"	0.000443984789603443
-"Rapeseed"	"phyllot"	"Default"	173.488950276243
-"Rapeseed"	"phyllot"	"q1"	127
-"Rapeseed"	"phyllot"	"q3"	230
