--- archive/data/parameters/type_params_fababean.csv
@@ -1,103 +1,0 @@
-"Species"	"Treatment"	"Parameter"	"Level"	"Value"
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_width_leaflet1"	"Default"	0.674369883186819
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_width_leaflet1"	"q1"	0.608143839238498
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_width_leaflet1"	"q3"	0.736314402810304
-"Fababean"	"intercrop_faba_vigo"	"coeff_width_leaflet1"	"Default"	0.739338834306975
-"Fababean"	"intercrop_faba_vigo"	"coeff_width_leaflet1"	"q1"	0.681427125506073
-"Fababean"	"intercrop_faba_vigo"	"coeff_width_leaflet1"	"q3"	0.765389876880985
-"Fababean"	"monocrop_faba"	"coeff_width_leaflet1"	"Default"	0.689338593883058
-"Fababean"	"monocrop_faba"	"coeff_width_leaflet1"	"q1"	0.651434426229508
-"Fababean"	"monocrop_faba"	"coeff_width_leaflet1"	"q3"	0.737533692722372
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_width_leaflet2"	"Default"	0.536716014505183
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_width_leaflet2"	"q1"	0.5
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_width_leaflet2"	"q3"	0.563741422896353
-"Fababean"	"intercrop_faba_vigo"	"coeff_width_leaflet2"	"Default"	0.467897535010638
-"Fababean"	"intercrop_faba_vigo"	"coeff_width_leaflet2"	"q1"	0.372549019607843
-"Fababean"	"intercrop_faba_vigo"	"coeff_width_leaflet2"	"q3"	0.552631578947368
-"Fababean"	"monocrop_faba"	"coeff_width_leaflet2"	"Default"	0.532413621331125
-"Fababean"	"monocrop_faba"	"coeff_width_leaflet2"	"q1"	0.5
-"Fababean"	"monocrop_faba"	"coeff_width_leaflet2"	"q3"	0.573770491803279
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_petiole_leaflet1"	"Default"	0.227539887431725
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_petiole_leaflet1"	"q1"	0.185984354628422
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_petiole_leaflet1"	"q3"	0.305397727272727
-"Fababean"	"intercrop_faba_vigo"	"coeff_petiole_leaflet1"	"Default"	0.19943638296351
-"Fababean"	"intercrop_faba_vigo"	"coeff_petiole_leaflet1"	"q1"	0.136363636363636
-"Fababean"	"intercrop_faba_vigo"	"coeff_petiole_leaflet1"	"q3"	0.288461538461538
-"Fababean"	"monocrop_faba"	"coeff_petiole_leaflet1"	"Default"	0.25021489346317
-"Fababean"	"monocrop_faba"	"coeff_petiole_leaflet1"	"q1"	0.176912568306011
-"Fababean"	"monocrop_faba"	"coeff_petiole_leaflet1"	"q3"	0.356501391348343
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_petiole_leaflet2"	"Default"	0.940215372841788
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_petiole_leaflet2"	"q1"	0.833333333333333
-"Fababean"	"intercrop_faba_aviso_RRF"	"coeff_petiole_leaflet2"	"q3"	1.04
-"Fababean"	"intercrop_faba_vigo"	"coeff_petiole_leaflet2"	"Default"	0.531815961585993
-"Fababean"	"intercrop_faba_vigo"	"coeff_petiole_leaflet2"	"q1"	0.197368421052632
-"Fababean"	"intercrop_faba_vigo"	"coeff_petiole_leaflet2"	"q3"	0.789473684210526
-"Fababean"	"monocrop_faba"	"coeff_petiole_leaflet2"	"Default"	0.997521238367122
-"Fababean"	"monocrop_faba"	"coeff_petiole_leaflet2"	"q1"	0.888544891640867
-"Fababean"	"monocrop_faba"	"coeff_petiole_leaflet2"	"q3"	1.11346153846154
-"Fababean"	"intercrop_faba_aviso_RRF"	"ins_angle"	"Default"	36.5909090909091
-"Fababean"	"intercrop_faba_aviso_RRF"	"ins_angle"	"q1"	25
-"Fababean"	"intercrop_faba_aviso_RRF"	"ins_angle"	"q3"	45
-"Fababean"	"intercrop_faba_vigo"	"ins_angle"	"Default"	52.5
-"Fababean"	"intercrop_faba_vigo"	"ins_angle"	"q1"	33.75
-"Fababean"	"intercrop_faba_vigo"	"ins_angle"	"q3"	70
-"Fababean"	"monocrop_faba"	"ins_angle"	"Default"	35.3378378378378
-"Fababean"	"monocrop_faba"	"ins_angle"	"q1"	30
-"Fababean"	"monocrop_faba"	"ins_angle"	"q3"	43.75
-"Fababean"	"intercrop_faba_aviso_RF"	"ins_angle"	"Default"	NA
-"Fababean"	"intercrop_faba_aviso_RF"	"ins_angle"	"q1"	NA
-"Fababean"	"intercrop_faba_aviso_RF"	"ins_angle"	"q3"	NA
-"Fababean"	"intercrop_faba_aviso_RRF"	"rmax"	"Default"	0.631174811075094
-"Fababean"	"intercrop_faba_aviso_RRF"	"rmax"	"q1"	0.565606867122757
-"Fababean"	"intercrop_faba_aviso_RRF"	"rmax"	"q3"	0.716862590850988
-"Fababean"	"intercrop_faba_vigo"	"rmax"	"Default"	0.703267425073356
-"Fababean"	"intercrop_faba_vigo"	"rmax"	"q1"	0.635581723665896
-"Fababean"	"intercrop_faba_vigo"	"rmax"	"q3"	0.76939753661012
-"Fababean"	"monocrop_faba"	"rmax"	"Default"	0.642321291818379
-"Fababean"	"monocrop_faba"	"rmax"	"q1"	0.564488613361871
-"Fababean"	"monocrop_faba"	"rmax"	"q3"	0.722340850504923
-"Fababean"	"intercrop_faba_aviso_RRF"	"k"	"Default"	2.32545671228301
-"Fababean"	"intercrop_faba_aviso_RRF"	"k"	"q1"	1.2262242541545
-"Fababean"	"intercrop_faba_aviso_RRF"	"k"	"q3"	2.11939966290699
-"Fababean"	"intercrop_faba_vigo"	"k"	"Default"	4.17543840243239
-"Fababean"	"intercrop_faba_vigo"	"k"	"q1"	1.05468649830538
-"Fababean"	"intercrop_faba_vigo"	"k"	"q3"	5.7823412220726
-"Fababean"	"monocrop_faba"	"k"	"Default"	2.16111930978653
-"Fababean"	"monocrop_faba"	"k"	"q1"	1.2283648538666
-"Fababean"	"monocrop_faba"	"k"	"q3"	2.20300122289185
-"Fababean"	"monocrop_faba"	"L_height"	"Default"	71.914691532791
-"Fababean"	"monocrop_faba"	"L_height"	"q1"	NA
-"Fababean"	"monocrop_faba"	"L_height"	"q3"	NA
-"Fababean"	"monocrop_faba"	"x0_height"	"Default"	537.427579096911
-"Fababean"	"monocrop_faba"	"x0_height"	"q1"	NA
-"Fababean"	"monocrop_faba"	"x0_height"	"q3"	NA
-"Fababean"	"monocrop_faba"	"k_height"	"Default"	0.0218146935545291
-"Fababean"	"monocrop_faba"	"k_height"	"q1"	NA
-"Fababean"	"monocrop_faba"	"k_height"	"q3"	NA
-"Fababean"	"intercrop_faba_aviso_RRF"	"L_height"	"Default"	64.7912294928082
-"Fababean"	"intercrop_faba_aviso_RRF"	"L_height"	"q1"	NA
-"Fababean"	"intercrop_faba_aviso_RRF"	"L_height"	"q3"	NA
-"Fababean"	"intercrop_faba_aviso_RRF"	"x0_height"	"Default"	521.41274635353
-"Fababean"	"intercrop_faba_aviso_RRF"	"x0_height"	"q1"	NA
-"Fababean"	"intercrop_faba_aviso_RRF"	"x0_height"	"q3"	NA
-"Fababean"	"intercrop_faba_aviso_RRF"	"k_height"	"Default"	0.0221059857423279
-"Fababean"	"intercrop_faba_aviso_RRF"	"k_height"	"q1"	NA
-"Fababean"	"intercrop_faba_aviso_RRF"	"k_height"	"q3"	NA
-"Fababean"	"intercrop_faba_vigo"	"L_height"	"Default"	55.2036327223325
-"Fababean"	"intercrop_faba_vigo"	"L_height"	"q1"	NA
-"Fababean"	"intercrop_faba_vigo"	"L_height"	"q3"	NA
-"Fababean"	"intercrop_faba_vigo"	"x0_height"	"Default"	506.084987781774
-"Fababean"	"intercrop_faba_vigo"	"x0_height"	"q1"	NA
-"Fababean"	"intercrop_faba_vigo"	"x0_height"	"q3"	NA
-"Fababean"	"intercrop_faba_vigo"	"k_height"	"Default"	0.0208841548828232
-"Fababean"	"intercrop_faba_vigo"	"k_height"	"q1"	NA
-"Fababean"	"intercrop_faba_vigo"	"k_height"	"q3"	NA
-"Fababean"	"intercrop_faba_vigo"	"phyllot"	"Default"	194.318181818182
-"Fababean"	"intercrop_faba_vigo"	"phyllot"	"q1"	156
-"Fababean"	"intercrop_faba_vigo"	"phyllot"	"q3"	231
-"Fababean"	"monocrop_faba"	"phyllot"	"Default"	170.176470588235
-"Fababean"	"monocrop_faba"	"phyllot"	"q1"	159.75
-"Fababean"	"monocrop_faba"	"phyllot"	"q3"	190.5
-"Fababean"	"intercrop_faba_aviso_RRF"	"phyllot"	"Default"	180.5
-"Fababean"	"intercrop_faba_aviso_RRF"	"phyllot"	"q1"	170.25
-"Fababean"	"intercrop_faba_aviso_RRF"	"phyllot"	"q3"	190.5
