--- archive/data/parameters/type_params_rapeseed.csv
@@ -1,88 +1,0 @@
-"Species"	"Treatment"	"Parameter"	"Level"	"Value"
-"Rapeseed"	"monocrop_aviso"	"coeff_width_leaf"	"Default"	0.432373902691464
-"Rapeseed"	"monocrop_aviso"	"coeff_width_leaf"	"q1"	0.391737443640558
-"Rapeseed"	"monocrop_aviso"	"coeff_width_leaf"	"q3"	0.478349944629014
-"Rapeseed"	"monocrop_vigo"	"coeff_width_leaf"	"Default"	0.393038642261872
-"Rapeseed"	"monocrop_vigo"	"coeff_width_leaf"	"q1"	0.337819111834457
-"Rapeseed"	"monocrop_vigo"	"coeff_width_leaf"	"q3"	0.439072164948454
-"Rapeseed"	"intercrop_aviso_RRF"	"coeff_width_leaf"	"Default"	0.432955729851061
-"Rapeseed"	"intercrop_aviso_RRF"	"coeff_width_leaf"	"q1"	0.389242492155984
-"Rapeseed"	"intercrop_aviso_RRF"	"coeff_width_leaf"	"q3"	0.471030222073802
-"Rapeseed"	"intercrop_vigo"	"coeff_width_leaf"	"Default"	0.377603085947721
-"Rapeseed"	"intercrop_vigo"	"coeff_width_leaf"	"q1"	0.318547657804305
-"Rapeseed"	"intercrop_vigo"	"coeff_width_leaf"	"q3"	0.429010538641686
-"Rapeseed"	"monocrop_aviso"	"coeff_petiole_leaf"	"Default"	0.395093952763011
-"Rapeseed"	"monocrop_aviso"	"coeff_petiole_leaf"	"q1"	0.344476055002371
-"Rapeseed"	"monocrop_aviso"	"coeff_petiole_leaf"	"q3"	0.442434379825013
-"Rapeseed"	"monocrop_vigo"	"coeff_petiole_leaf"	"Default"	0.397661436124243
-"Rapeseed"	"monocrop_vigo"	"coeff_petiole_leaf"	"q1"	0.372746253991648
-"Rapeseed"	"monocrop_vigo"	"coeff_petiole_leaf"	"q3"	0.43857238810527
-"Rapeseed"	"intercrop_aviso_RRF"	"coeff_petiole_leaf"	"Default"	0.36244952462683
-"Rapeseed"	"intercrop_aviso_RRF"	"coeff_petiole_leaf"	"q1"	0.284115138592751
-"Rapeseed"	"intercrop_aviso_RRF"	"coeff_petiole_leaf"	"q3"	0.430994300342945
-"Rapeseed"	"intercrop_vigo"	"coeff_petiole_leaf"	"Default"	0.249667529747796
-"Rapeseed"	"intercrop_vigo"	"coeff_petiole_leaf"	"q1"	0.14691706469822
-"Rapeseed"	"intercrop_vigo"	"coeff_petiole_leaf"	"q3"	0.323293172690763
-"Rapeseed"	"monocrop_aviso"	"ins_angle"	"Default"	33.8461538461538
-"Rapeseed"	"monocrop_aviso"	"ins_angle"	"q1"	20
-"Rapeseed"	"monocrop_aviso"	"ins_angle"	"q3"	35
-"Rapeseed"	"monocrop_vigo"	"ins_angle"	"Default"	44.0909090909091
-"Rapeseed"	"monocrop_vigo"	"ins_angle"	"q1"	26.25
-"Rapeseed"	"monocrop_vigo"	"ins_angle"	"q3"	57.5
-"Rapeseed"	"intercrop_aviso_RRF"	"ins_angle"	"Default"	33.4097222222222
-"Rapeseed"	"intercrop_aviso_RRF"	"ins_angle"	"q1"	20
-"Rapeseed"	"intercrop_aviso_RRF"	"ins_angle"	"q3"	40
-"Rapeseed"	"intercrop_vigo"	"ins_angle"	"Default"	31.969696969697
-"Rapeseed"	"intercrop_vigo"	"ins_angle"	"q1"	10
-"Rapeseed"	"intercrop_vigo"	"ins_angle"	"q3"	45
-"Rapeseed"	"monocrop_aviso"	"rmax"	"Default"	0.624657902692764
-"Rapeseed"	"monocrop_aviso"	"rmax"	"q1"	0.573486047966061
-"Rapeseed"	"monocrop_aviso"	"rmax"	"q3"	0.648065096115762
-"Rapeseed"	"monocrop_vigo"	"rmax"	"Default"	0.60905259691865
-"Rapeseed"	"monocrop_vigo"	"rmax"	"q1"	0.549124220262684
-"Rapeseed"	"monocrop_vigo"	"rmax"	"q3"	0.667733659045733
-"Rapeseed"	"intercrop_aviso_RRF"	"rmax"	"Default"	0.624251194032368
-"Rapeseed"	"intercrop_aviso_RRF"	"rmax"	"q1"	0.57471342414942
-"Rapeseed"	"intercrop_aviso_RRF"	"rmax"	"q3"	0.655703268653379
-"Rapeseed"	"intercrop_vigo"	"rmax"	"Default"	0.609686904351311
-"Rapeseed"	"intercrop_vigo"	"rmax"	"q1"	0.519870144010147
-"Rapeseed"	"intercrop_vigo"	"rmax"	"q3"	0.70569995511487
-"Rapeseed"	"monocrop_aviso"	"k"	"Default"	6.49561736248202
-"Rapeseed"	"monocrop_aviso"	"k"	"q1"	3.62740714828621
-"Rapeseed"	"monocrop_aviso"	"k"	"q3"	8.52980046109038
-"Rapeseed"	"monocrop_vigo"	"k"	"Default"	5.60088036045245
-"Rapeseed"	"monocrop_vigo"	"k"	"q1"	2.72717725196216
-"Rapeseed"	"monocrop_vigo"	"k"	"q3"	8.24314708062631
-"Rapeseed"	"intercrop_aviso_RRF"	"k"	"Default"	7.39074533273133
-"Rapeseed"	"intercrop_aviso_RRF"	"k"	"q1"	3.28391734453831
-"Rapeseed"	"intercrop_aviso_RRF"	"k"	"q3"	10.5782617431435
-"Rapeseed"	"intercrop_vigo"	"k"	"Default"	9.96232011705795
-"Rapeseed"	"intercrop_vigo"	"k"	"q1"	4.0252427278207
-"Rapeseed"	"intercrop_vigo"	"k"	"q3"	10.8676893812091
-"Rapeseed"	"monocrop_aviso"	"growth_node"	"Default"	0.000254573569537356
-"Rapeseed"	"monocrop_aviso"	"growth_node"	"q1"	NA
-"Rapeseed"	"monocrop_aviso"	"growth_node"	"q3"	NA
-"Rapeseed"	"monocrop_vigo"	"growth_node"	"Default"	0.000373880555395238
-"Rapeseed"	"monocrop_vigo"	"growth_node"	"q1"	NA
-"Rapeseed"	"monocrop_vigo"	"growth_node"	"q3"	NA
-"Rapeseed"	"intercrop_aviso_RRF"	"growth_node"	"Default"	0.000266610548751953
-"Rapeseed"	"intercrop_aviso_RRF"	"growth_node"	"q1"	NA
-"Rapeseed"	"intercrop_aviso_RRF"	"growth_node"	"q3"	NA
-"Rapeseed"	"intercrop_vigo"	"growth_node"	"Default"	0.000501045428631635
-"Rapeseed"	"intercrop_vigo"	"growth_node"	"q1"	NA
-"Rapeseed"	"intercrop_vigo"	"growth_node"	"q3"	NA
-"Rapeseed"	"intercrop_aviso_RF"	"growth_node"	"Default"	0.000161864508188314
-"Rapeseed"	"intercrop_aviso_RF"	"growth_node"	"q1"	NA
-"Rapeseed"	"intercrop_aviso_RF"	"growth_node"	"q3"	NA
-"Rapeseed"	"monocrop_vigo"	"phyllot"	"Default"	170.714285714286
-"Rapeseed"	"monocrop_vigo"	"phyllot"	"q1"	131
-"Rapeseed"	"monocrop_vigo"	"phyllot"	"q3"	213.25
-"Rapeseed"	"intercrop_vigo"	"phyllot"	"Default"	168.734042553192
-"Rapeseed"	"intercrop_vigo"	"phyllot"	"q1"	114
-"Rapeseed"	"intercrop_vigo"	"phyllot"	"q3"	215
-"Rapeseed"	"intercrop_aviso_RRF"	"phyllot"	"Default"	186.0625
-"Rapeseed"	"intercrop_aviso_RRF"	"phyllot"	"q1"	142.125
-"Rapeseed"	"intercrop_aviso_RRF"	"phyllot"	"q3"	244
-"Rapeseed"	"monocrop_aviso"	"phyllot"	"Default"	167.5
-"Rapeseed"	"monocrop_aviso"	"phyllot"	"q1"	123
-"Rapeseed"	"monocrop_aviso"	"phyllot"	"q3"	219.25
